api/pyproject.toml
[tool.poetry]
name = "social-network"
version = "0.1.0"
description = ""
authors = ["pajecawav <pajecawav@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
pydantic = {extras = ["dotenv"], version = "^1.8.1"}
uvicorn = {extras = ["standard"], version = "^0.13.4"}
SQLAlchemy = "^1.4.8"
python-jose = "^3.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
psycopg2-binary = "^2.8.6"
alembic = "^1.5.8"
python-socketio = "^5.2.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
flake8 = "^3.9.1"
mypy = "^0.812"
sqlalchemy-stubs = "^0.4"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
